Add "Admin" Cloud Function.
Move the UpdateRoutes Cloud Function into a single Admin
function that can later be updated support additional
functionality.

To make it possible to deploy the function from the root
project directory, the main entry point into the function is
now an ascenso.go file, with its implementation spread
across multiple files in a new admin package.
